sql/catalog/database: extract package for the DatabaseCache
Release note: None
