+++ pkg/sql/database_test.go
@@ -20,32 +20,11 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/keys"
 	"github.com/cockroachdb/cockroach/pkg/kv"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/catalogkv"
-	"github.com/cockroachdb/cockroach/pkg/sql/parser"
-	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
+	"github.com/cockroachdb/cockroach/pkg/sql/catalog/database"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlbase"
 	"github.com/cockroachdb/cockroach/pkg/testutils/serverutils"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
 )
-
-func TestMakeDatabaseDesc(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-
-	stmt, err := parser.ParseOne("CREATE DATABASE test")
-	if err != nil {
-		t.Fatal(err)
-	}
-	desc := MakeDatabaseDesc(stmt.AST.(*tree.CreateDatabase))
-	if desc.Name != "test" {
-		t.Fatalf("expected Name == test, got %s", desc.Name)
-	}
-	// ID is not set yet.
-	if desc.ID != 0 {
-		t.Fatalf("expected ID == 0, got %d", desc.ID)
-	}
-	if len(desc.GetPrivileges().Users) != 2 {
-		t.Fatalf("wrong number of privilege users, expected 2, got: %d", len(desc.GetPrivileges().Users))
-	}
-}
 
 func TestDatabaseAccessors(t *testing.T) {
 	defer leaktest.AfterTest(t)()
@@ -61,7 +40,7 @@ func TestDatabaseAccessors(t *testing.T) {
 			return err
 		}
 
-		databaseCache := NewDatabaseCache(keys.SystemSQLCodec, config.NewSystemConfig(zonepb.DefaultZoneConfigRef()))
+		databaseCache := database.NewDatabaseCache(keys.SystemSQLCodec, config.NewSystemConfig(zonepb.DefaultZoneConfigRef()))
 		_, err := databaseCache.GetDatabaseDescByID(ctx, txn, sqlbase.SystemDB.ID)
 		return err
 	}); err != nil {
